Sprint#3 User can now edit a task
We can now edit a task when we press on a task. If
the edited task contains a title that already exists,
an error messsage is displayed and the user can't update the task.
Sprint#3 User can now edit a task

We can now edit a task when we press on a task. If
the edited task contains a title that already exists,
an error messsage is displayed and the user can't update the task.
So we make sure that a title is unique.