Editing a task is now tested
User can now edit a task by long clicking on a task
in the main activity and click on the edit button.
The JavaDoc and the tests are now done. They check that when
a user edit a task, the task is updated in the local database
and the updated version is shown in the main activity.
When the user edit a task, he cannot enter an empty
title or a title that is already used by another task.
(This will be useful to ensure that title are unique in the database).

Note: the test coverage is currently at 91% for the project.