Create a listView to display the task's information
In the TaskInformationActivity, the task's informations are displayed by a listView.
Each task's information is represented by a InformationItem which is constituted by the title on the information and the body.
The title is the information type. It can be "Title" or "Description" etc...
The body is the information. For example it can be the task's title or the task's description.