#52 reimplementation of login test and removeAccount creation
The test for google login using UIAutomator have been reimplemented and are being revised in order to make them work to increase test coverage. An utilitary tearDown function, removeAccount, has also been created in order to remove the account after the login to be sure not to compromise future rerun of the login test.