Implementation of notifications
Create notification to notifiy the user with some delay before
the due date according to the needed time of the task. If the task
take an important amount of time, the user will be notified earlier.
Notifications are also updated when a task is update or removed,
these operations are done in a background thread in order to stacking
operations on the UI thread.

Tests still need to be implemented though.